backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
telethon==1.30.3
asyncpg==0.28.0
bcrypt==4.0.1
alembic==1.12.1
httpx==0.25.1 